build: import the AM_SILENT_RULES if available, but default to verbose.
Automake 1.11 (and I think 1.10b already) offers the AM_SILENT_RULES macro.
This adds switches --quiet, --enable-silent-rules and --disable-silent-rules
to configure.

--quiet makes the configure run itself quite.
--enable-silent-rules makes the compile process less verbose:
  for a file that is compiled without errors or warnings, a simple
  "CC     main.o" is printed (e.g.). Compiler warnings and errors
  are printed of course.
  This makes it much easier (IMHO) to spot build problems.
--disable-silent-rules turns the silent rules off

I have set it up such that the default for tinyproxy is to build
in verbose mode (i.e. with silent rules disabled). This prints
the whole compile call command line for each source file compiled,
precisely as before.

You can also control verbose/non-verbose mode at "make" time, i.e.
after configure has run, by calling "make V=0 ..." or "make V=1 ..."
for running in silent and verbose mode, respectively.

If the version automake used to create configure is too old,
the result is unaltered, compared to the result before this change.

Wow - this is a long commit message for a 1-liner.
But since I discussed this with Mukund earlier, and he did
not seem to be too fond if this, I felt the need to justify
this change... :-)

Michael