do not daemonize by default, change -d switch to -D
having the program forking into background by default is a very
unintuitive behaviour, and in case an error happens the process
terminates "silently" (that is, it will only log the error to logfile).
rather than reversing the meaning of the -d command line option, we
replace it with -D so users will notice the change and can change their
scripts, instead of getting a silently different behaviour.