Add support for the PROXY protocol
When there is a proxy in front of tinyproxy, the latter does not see the
actual IP address of the client. This is a problem when using tools like
fail2ban that parse the tinyproxy log to find and ban clients that try
to break the authentication using brute force. Some proxies like stunnel
or HAProxy support the PROXY protocol to overcome that issue, so that
the next proxy in the chain can tell what the IP address of the original
client is.

See https://www.haproxy.org/download/1.8/doc/proxy-protocol.txt

This commit adds support for clients that use the PROXY protocol to
extract the real client IP address and report it in the logs. This is
only implemented when the new ClientUsesProxyProtocol parameter is set
to Yes. It should only be enabled when the client is trusted and known
to be a proxy that support that protocol.